Update README.md
I noticed you have the official Sapper template here, so I hope it's okay I added an updated version with tailwindcss support.

I think it would be helpful for anyone looking to implement tailwindcss as this is the "goto" place for all things Svelte. Personally I looked through all the Svelte-tailwindcss templates in vain hoping for help to implement it in Sapper.


And thanks for a great list! 🍺